Move giant actor switch statement to Actor.ts
